requirements.txt
pwntools
requests
pytest-shutil
python-libarchive